Adopt minor suggestions from #14
Refs https://github.com/greenelab/scihub-manuscript/issues/14

Addresses comment from @rossmounce:

I may be wrong but to me 'fate' can include or imply death and that's a bit much.
"which will persist regardless of the fate of Sci-Hub" instead?

Addresses comment from @tamunro:

On the phrase in the abstract "the overwhelming majority of scholarly
literature", I think the "nearly all" in the title is punchier, and
true; if the intention is to distinguish access from archiving, surely
this can be left to the body of the paper.